Feat: add logic to display a user's photos
Created a static_pages controller and index view with its action to
handle this feature. Also added the route in the routes file.

In the action, there is a method that creates a flickr object, then it
calls a method to obtain the response with the photos of a user using
the id. Then with a loop, assigning the server, id and secret attributes
to an url provided by flickr to generate each photos link. Store them in
and instance variable.

In the view, there is a form to enter a user id and a submit button.
When the user submits the id and if the searched user contains photos,
it uses a loop and an image_tag with a link to each photo.